move handleMobilePress to utils
