+++ src/components/Message/hooks/useBreakpoint.js
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import throttle from 'lodash.throttle';
 
+/** @type {(width: number) => {device: 'mobile' | 'tablet' | 'full'; width: number}} */
 const getDeviceWidth = (width) => {
   if (width < 768) return { device: 'mobile', width };
   if (width < 1024) return { device: 'tablet', width };
